% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBMR.R
\name{runshinyBMR}
\alias{runshinyBMR}
\title{Run shinyBMR}
\usage{
runshinyBMR(...)
}
\arguments{
\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyBMR.
}
\examples{
\dontrun{
  runShinyBMR()
}
}
